feat: added semaphore when test is running
